moved thumbnail to its own component
